Refactored costs in db module

+++ db/src/main/sql/create-tables.sql
@@ -274,11 +274,11 @@ CREATE TABLE `carcosts` (
 	`car_cost_id` INT NOT NULL AUTO_INCREMENT,
 	`car_cost_car_id` INT NOT NULL,
 	`car_cost_proof` INT,
-	`car_cost_amount` DECIMAL(19,4) NOT NULL,
+	`car_cost_amount` INT NOT NULL,
 	`car_cost_description` TEXT,
 	`car_cost_status` ENUM('REQUEST','ACCEPTED', 'REFUSED', 'FROZEN') NOT NULL DEFAULT 'REQUEST', -- approved by car_admin
 	`car_cost_time` DATE,
-	`car_cost_mileage` DECIMAL(10,1),
+	`car_cost_mileage` INT,
 	`car_cost_billed` DATE DEFAULT NULL,
 	`car_cost_created_at` TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
 	`car_cost_updated_at` TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
